code/requirements.txt
anthropic>=0.49.0
typer[all]>=0.12.0
rich>=13.7.0
rank-bm25>=0.2.2
pydantic>=2.7.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
pytest>=8.2.0
pytest-cov>=5.0.0
